+++ changelog.md
@@ -6,6 +6,10 @@ Removed a load of unused strings
 Added extension date to report
 Added due date to 'export all' report
 
-14
+1.2.4
 Tested (and working) with 3.5 and 3.6
 No functional changes
+
+1.2.5
+Fixed bug where grader is shown instead of marker
+Show marking workflow, grader and marker in report